refactor(context-menu): H.5 — extract MultiArtistToPlaylistSubmenu
